Add "Ask Question" feature in the FAQ page
feat: add the FAQ page menu in the navbar of both "index.html" and "FAQ.hml" pages

fix: the navbar anchors in the FAQ page are now redirecting the user correctly.

the anchors were redirecting the user to nothing, since the hrefs were set to redirect to elements that are actually in the index page. I fixed it, and now the anchors are redirecting the user first to the index page and then to an element. It's working fine now